requirements.txt
# generated from manifests external_dependencies
chardet
ofxparse
ofxtools
plaid-python
xlrd
